Keeping JSONs...
There are some spots in some sites that complain about the frozen hashes
or something. This kinda sounds like we're trying to mutate something
that should not be mutated, and there's a part of me that wants to
investigate that instead.

But for now, reverting the JSON-related changes gets the site to build.