Removing JSONs, again
Yeah so the errors we were seeing WAS a problem with the site. Using
JSONs here "fixed" it in the sense that the site built, but lost all the
locale magic.

I'm reverting the reverted changes. It's arguably cleaner, and at least
we'll get a (cryptic) warning that something is off in the future.