Created the ability to support namespaced queries.
This was needed incase the user put the queries in a subfolder.  It will
pull the last class name off and use that as the query name.

So Queries::MyModule::Profile will expect the datocms model to be named
"Profile".

Same with the rails intergation, the model will still be "Profile" none
of the other stuff matters at this time.

Also ran into an issue in Bridgetown where the classes were being
re-created over and over again.  I put in a unique clause into the
search for query names so that it would clear things up.

Just realized that this may cause a problem.... Investigating now.