Optional API key
Only parsing queries if the client exists.

The point is that the client exists only if it _can_ exist -- i.e. if
we're in test or we have a proper API Key.

If it doesn't we're still getting the log message about missing keys,
and `parse` won't do anything. This obviously means that nothing will
work, but it will only break stuff when we actually try to use the query
results rather than at start up.

So, other things that don't really care about the data (e.g. rake tasks
or whatever) will still start and be fine as long as they don't use the
Graph models.